fix: separate the manager of the insight and recommendation processing part of the recommender service
