+++ src/spaceone/inventory/conf/cloud_service_conf.py
@@ -40,7 +40,7 @@ CLOUD_SERVICE_GROUP_MAP = {
         'FunctionGen1Manager'
     ],
     'Recommender': [
-        'RecommenderManager',
+        'InsightManager'
     ]
 }
