+++ src/spaceone/inventory/manager/recommender/insight_manager.py
@@ -5,23 +5,20 @@ import requests
 from bs4 import BeautifulSoup
 
 from spaceone.inventory.conf.cloud_service_conf import REGION_INFO
-from spaceone.inventory.connector import RecommendationConnector
 from spaceone.inventory.libs.manager import GoogleCloudManager
 from spaceone.inventory.libs.schema.base import ReferenceModel
 from spaceone.inventory.connector.recommender.insight import InsightConnector
 from spaceone.inventory.connector.recommender.cloud_asset import CloudAssetConnector
 from spaceone.inventory.model.recommender.insight.cloud_service import InsightResource, InsightResponse
-from spaceone.inventory.model.recommender.recommendation.cloud_service import RecommendationResource, \
-    RecommendationResponse
 from spaceone.inventory.model.recommender.insight.cloud_service_type import CLOUD_SERVICE_TYPES
 from spaceone.inventory.model.recommender.insight.data import Insight, IPAddressInsight
-from spaceone.inventory.model.recommender.recommendation.data import Recommendation
-from spaceone.inventory.model.recommender.recommendation.cloud_sevice_type import RECOMMENDATION_CLOUD_SERVICE_TYPES
+from spaceone.inventory.manager.recommender.recommendation_manager import RecommendationManager
+
 
 _LOGGER = logging.getLogger(__name__)
 
 
-class RecommenderManager(GoogleCloudManager):
+class InsightManager(GoogleCloudManager):
     connector_name = 'InsightConnector'
     cloud_service_types = CLOUD_SERVICE_TYPES
 
@@ -76,21 +73,11 @@ class RecommenderManager(GoogleCloudManager):
                                                                            origin_insight.get('name'))
                     error_responses.append(error_response)
 
-        recommendation_conn: RecommendationConnector = self.locator.get_connector(RecommendationConnector, **params)
-        self.cloud_service_types = RECOMMENDATION_CLOUD_SERVICE_TYPES
-        for recommendation_name in target_recommendation_names:
-            try:
-                recommendation = recommendation_conn.get_recommendation(recommendation_name)
-                recommendation_response = self._create_recommendation_response(recommendation, project_id)
-                collected_cloud_services.append(recommendation_response)
-
-            except Exception as e:
-                _LOGGER.error(f'[collect_cloud_service] => {e}', exc_info=True)
-                error_response = self.generate_resource_error_response(e, 'Recommender', 'Recommendation',
-                                                                       recommendation_name)
-                error_responses.append(error_response)
-
         _LOGGER.debug(f'** Recommender Insight Finished {time.time() - start_time} Seconds **')
+
+        params['recommendation_names'] = target_recommendation_names
+        RecommendationManager().collect_cloud_service(params)
+
         return collected_cloud_services, error_responses
 
     def _create_insight_type_by_crawling(self):
@@ -197,44 +184,9 @@ class RecommenderManager(GoogleCloudManager):
         })
         return InsightResponse({'resource': insight_resource})
 
-    def _create_recommendation_response(self, recommendation, project_id):
-        region, instance_type = self._get_region_and_instance_type(recommendation['name'])
-
-        display = {
-            'instance_type': instance_type
-        }
-
-        recommendation.update({
-            'display': display
-        })
-
-        recommendation_data = Recommendation(recommendation, strict=False)
-        recommendation_resource = RecommendationResource({
-            'name': recommendation_data.get('name', 'Unknown'),
-            'account': project_id,
-            'tags': {},
-            'region_code': region,
-            'instance_type': '',
-            'instance_size': 0,
-            'data': recommendation_data,
-            'reference': ReferenceModel(recommendation_data.reference())
-        })
-
-        return RecommendationResponse({'resource': recommendation_resource})
-
     @staticmethod
     def _list_recommendation_names(recommendations):
         recommendation_names = []
         for recommendation in recommendations:
             recommendation_names.append(recommendation['recommendation'])
         return recommendation_names
-
-    @staticmethod
-    def _get_region_and_instance_type(name):
-        try:
-            project_id, resource = name.split('locations/')
-            region, _, instance_type, _ = resource.split('/', 3)
-            return region, instance_type
-
-        except Exception as e:
-            _LOGGER.error(f'[_get_region] recommendation passing error (data: {name}) => {e}', exc_info=True)
